Modernise test_coord_categorisation.
